Add minimal SSZ reference testing for v0.8.1 - Eth1Data
